requirements.txt
# Dependencies

aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.1
distro==1.8.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langsmith==0.0.62
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.2.1
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-subtests==0.11.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.1
distro==1.8.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langsmith==0.0.62
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.2.1
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
dataclasses-json==0.6.1
distro==1.8.0
frozenlist==1.4.0
google-search-results==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.333
langsmith==0.0.62
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==1.2.1
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.3
pytest-subtests==0.11.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
